move WebsocketFrame to its own file
