+++ lib/websocket/websocket.js
@@ -15,7 +15,8 @@ const {
   kResponse
 } = require('./symbols')
 const { isEstablished, isClosing } = require('./util')
-const { establishWebSocketConnection, failWebsocketConnection, WebsocketFrame } = require('./connection')
+const { establishWebSocketConnection, failWebsocketConnection } = require('./connection')
+const { WebsocketFrame } = require('./frame')
 const { kEnumerableProperty, isBlobLike } = require('../core/util')
 const { types } = require('util')
